requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.28.0
attrs==23.1.0
bcrypt==4.1.2
billiard==4.2.0
celery==5.3.4
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.2.0
cryptography==41.0.7
cycler==0.12.1
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.105.0
fastapi-users-db-sqlalchemy==6.0.1
fonttools==4.47.0
frozenlist==1.4.0
greenlet==3.0.0
h11==0.14.0
idna==3.4
kiwisolver==1.4.5
kombu==5.3.3
makefun==1.15.2
matplotlib==3.8.2
multidict==6.0.4
numpy==1.26.2
packaging==23.2
passlib==1.7.4
Pillow==10.1.0
pip-autoremove==0.10.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
schedule==1.2.1
scipy==1.11.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.22
starlette==0.27.0
typing_extensions==4.8.0
tzdata==2023.3
uvicorn==0.23.2
vine==5.1.0
wcwidth==0.2.9
yarl==1.9.2
